--- src/main/java/com/healthshop/healthshop/MemberRepository.java
@@ -1,22 +1,0 @@
-package com.healthshop.healthshop;
-
-import com.healthshop.healthshop.domain.member.Member;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class MemberRepository {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    public Long save(Member member) {
-        em.persist(member);
-        return member.getId();
-    }
-
-    public Member find(Long id) {
-        return em.find(Member.class, id);
-    }
-}
