--- src/test/java/com/healthshop/healthshop/MemberRepositoryTest.java
@@ -1,38 +1,0 @@
-package com.healthshop.healthshop;
-
-import com.healthshop.healthshop.domain.member.Member;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.Rollback;
-import org.springframework.transaction.annotation.Transactional;
-
-import static org.assertj.core.api.Assertions.*;
-
-
-@SpringBootTest
-class MemberRepositoryTest {
-
-    @Autowired
-    MemberRepository memberRepository;
-
-    @Test
-    @Transactional
-    @Rollback(false)
-    public void testMember() throws Exception {
-        //given
-        Member member = new Member();
-        member.setUsername("memberA");
-
-        //when
-        Long savedId = memberRepository.save(member);
-        Member findMember = memberRepository.find(savedId);
-
-        //then
-        assertThat(findMember.getId()).isEqualTo(member.getId());
-        assertThat(findMember.getUsername()).isEqualTo(member.getUsername());
-        assertThat(findMember).isEqualTo(member);
-        System.out.println("findMember == member: " + (findMember == member));
-    }
-
-}
